// src/app/models/communication.model.ts
/**
 * Modèle de communication
 */
export interface Communication {

  id: number;

  title: string;

  type: string;

  description: string;

  report: string;

  eventDate: string;

  createdAt: string;

  documentName: string;

  documentUrl: string;

  documentType: string;

  accessRole: string;
}